# Optimize BGR Mat to vpImage<vpRGBa> conversion