Note:

- flip case is not handled since it should a marginal usage
- `CV_8UC4` is not handled since most of the time `CV_8UC3` format should be used

Warning:
- with optimized code, alpha channel is set to zero, without it is set to 255
- since alpha channel should almost be never used, I think it is OK

---

Benchmark code:
```
    rs2::pipeline pipe;
    rs2::config cfg;
//    const int w = 1920, h = 1080;
    const int w = 640, h = 480;
    cfg.enable_stream(RS2_STREAM_COLOR, w, h, RS2_FORMAT_BGR8);
    pipe.start(cfg);

    vpImage<vpRGBa> I(h,w);
    vpDisplayX d;
    d.init(I, 0, 0, "RealSense");

    std::vector<double> times;
    while (true)
    {
      rs2::frameset data = pipe.wait_for_frames(); // Wait for next set of frames from the camera
      rs2::frame color = data.get_color_frame();

      cv::Mat image(h, w, CV_8UC3, (void*)color.get_data());

      double t = vpTime::measureTimeMs();
      vpImageConvert::convert(image, I);
      t = vpTime::measureTimeMs() - t;
      times.push_back(t);

      vpDisplay::display(I);
      vpDisplay::flush(I);
      if (vpDisplay::getClick(I, false)) {
        break;
      }
    }

    if (!times.empty()) {
      std::cout << "Mean: " << vpMath::getMean(times)
                << " ms ; Median: " << vpMath::getMedian(times)
                << " ms ; Std: " << vpMath::getStdev(times)
                << std::endl;
    }
```

---

|                     | 640x480 (old) | 640x480 (new) | x-factor | 1920x1080 (old) | 1920x1080 (new) | x-factor
| --------------   | -------------        | -------------         | ---------- | -------------           | ----------------------- | ----------
| Mean (ms)   | 2.37779          | 0.21794            | 10.9      | 12.3351             | 1.62221                | 7.6
| Median(ms) | 2.36108          | 0.174072          |              | 11.033                | 1.34399                |
| Std (ms)      | 0.373192        | 0.128349          |              | 2.90749              | 0.758989              |